Moved TypeExtensions to Utils folder
