+++ src/Transport/Topology/PublishersConfiguration.cs
@@ -52,21 +52,4 @@
             return _publishers.ContainsKey(type);
         }
     }
-
-    internal static class TypeExtensions
-    {
-        public static IEnumerable<Type> GetParentTypes(this Type target)
-        {
-            foreach (var i in target.GetInterfaces())
-                yield return i;
-
-            var currentBaseType = target.BaseType;
-            var objectType = typeof(object);
-            while (currentBaseType != null && currentBaseType != objectType)
-            {
-                yield return currentBaseType;
-                currentBaseType = currentBaseType.BaseType;
-            }
-        }
-    }
 }
